requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap4==2025.6
cryptography==45.0.4
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
